Move Quicktune code to util/
